@@ -14,7 +14,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<!--<version>2.0.0.RELEASE</version>-->
+		<!--<version>2.6.7</version>-->
 		<version>1.5.10.RELEASE</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
@@ -22,8 +22,8 @@
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<java.version>1.8</java.version>
-		<!--<spring-cloud.version>Finchley.M8</spring-cloud.version>-->
+		<java.version>11</java.version>
+		<!--<spring-cloud.version>2021.0.2</spring-cloud.version>-->
         <spring-cloud.version>Edgware.RELEASE</spring-cloud.version>
 	</properties>
 
