@@ -2,9 +2,9 @@ package com.sivalabs.orderservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker;
+//import org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker;
 
-@EnableCircuitBreaker
+//@EnableCircuitBreaker
 @SpringBootApplication
 public class OrderServiceApplication {
 
