@@ -1,13 +1,13 @@
 package com.sivalabs.shoppingcartui;
 
-import com.sivalabs.shoppingcartui.filters.AuthHeaderFilter;
+//import com.sivalabs.shoppingcartui.filters.AuthHeaderFilter;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 //import org.springframework.boot.autoconfigure.security.oauth2.client.EnableOAuth2Sso;
-import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
+//import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
 import org.springframework.context.annotation.Bean;
 
-@EnableZuulProxy
+//@EnableZuulProxy
 //@EnableOAuth2Sso
 @SpringBootApplication
 public class ShoppingcartUiApplication {
@@ -16,8 +16,8 @@ public class ShoppingcartUiApplication {
         SpringApplication.run(ShoppingcartUiApplication.class, args);
     }
 
-    @Bean
+    /*@Bean
     AuthHeaderFilter authHeaderFilter() {
         return new AuthHeaderFilter();
-    }
+    }*/
 }
